separated helper functions to lexiconHelpers
